Support in story-viz for generating images via a button
Add backend API and frontend invocation of the API via a button for
each page in a story. Also add a carousel viewer for pages with more
than 4 images.
