+++ story-viz/frontend/src/App.jsx
@@ -4,7 +4,7 @@ import Login from './Login'
 import Metadata from './Metadata';
 import Pages from './Pages';
 import StoryList from './StoryList';
-import { fetchSelectionsData, fetchStoryData } from './api';
+import { fetchSelectionsData, fetchStoryData, postGenerateImages } from './api';
 import { useError } from './ErrorContext';
 import { useToken } from './TokenContext';
 import ErrorMessage from './ErrorMessage';
@@ -34,6 +34,24 @@ const Story = () => {
     }
   }, [storyId]);
 
+  const handleGenereateImages = async (pageIndex) => {
+    const isPayloadValid = Number.isInteger(pageIndex) &&
+                           pageIndex >= 0 &&
+                           pageIndex < data.pages.length;
+
+    if (!isPayloadValid) {
+      setError('Invalid payload: Page index must be a valid integer within the range of pages.');
+      return;
+    }
+    try {
+      await postGenerateImages(storyId, pageIndex, token);
+      const storyData = await fetchStoryData(storyId, token)
+      setData(storyData);
+    } catch (error) {
+      setError('Error generating images:', error);
+    }
+  }
+
   if (!data || selections === null) {
     return <div>Loading...</div>;
   }
@@ -42,7 +60,7 @@ const Story = () => {
     <div>
         <ErrorMessage />
         <Metadata data={data} storyUrl={storyUrl} />
-        <Pages pages={data.pages} selections={selections} storyId={storyId} />
+        <Pages pages={data.pages} selections={selections} storyId={storyId} handleGenereateImages={handleGenereateImages}/>
     </div>
   );
 };
