+++ story-viz/frontend/src/Pages.module.css
@@ -11,30 +11,3 @@
     padding: 10px;
     margin-bottom: 20px;
 }
-
-.imageContainer {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.imageWrapper {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin: 0 10px;
-}
-
-.image {
-    max-width: 100%;
-    cursor: pointer;
-    margin-bottom: 10px;
-}
-
-.selectedImage {
-    border: 3px solid blue;
-}
-
-.unselectedImage {
-    border: 3px solid #ffffff00;
-}
