+++ story-viz/requirements.txt
@@ -5,15 +5,20 @@ fastapi
 google-cloud-secret-manager
 gunicorn
 mypy
+openai
 passlib[bcrypt]
+Pillow
 pyflakes
 pyjwt
 pymongo[srv]
 pytest
 python-dotenv
 python-multipart
+replicate
 requests
 types-passlib
+types-Pillow
 types-requests
 uvicorn
+wandb
 watchdog
